--- a/apps/ledger-live-mobile/src/components/RootNavigator/LazyNavigator/BaseNavigator.lazy.ts
+++ b/apps/ledger-live-mobile/src/components/RootNavigator/LazyNavigator/BaseNavigator.lazy.ts
@@ -1,0 +1,9 @@
+import { NavigatorName } from "~/const";
+import { registerAppScreen } from "LLM/performance/apis";
+
+export const SettingsNavigator = registerAppScreen<
+  typeof import("~/components/RootNavigator/SettingsNavigator").default
+>({
+  loader: () => import("~/components/RootNavigator/SettingsNavigator"),
+  name: NavigatorName.Settings,
+});