… apps/ledger-live-mobile/src/components/RootNavigator/index.tsx
… apps/ledger-live-mobile/tsconfig.json
… pnpm-lock.yaml